:computer: Decouples search video handlers from show_live
